Rewrite test-tree generation in Python
This is essentially a fresh start.  I originally thought we might
re-use the list of special characters from testTree.php, but then I
realized it was better to just loop over the potentially troublemaking
regions of ASCII and get all the special characters that way.

I also changed the license to MIT since Permanent's upstream code is
under that.  There will likely be non-trivial bits of code in this
repository, so we might as well use a real FOSS license.  If we later
also have a non-trivial amount of raw data (e.g., sample source data)
then we could put the data under CC0 while keeping the code under MIT.